backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
openai==1.35.3
faiss-cpu==1.8.0
numpy==1.26.4
python-dotenv==1.0.1
gunicorn==22.0.0
